+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Development
 
+* Abort processing if the project has a `Gemfile` but no `Gemfile.lock`.
+
 ## v3.0.4
 
 * Fix an issue where `obj = Klass.any_instance; obj.should_receive(:message)` was not properly converted to `expect_any_instance_of(Klass).to receive(:message)` when run on RSpec 3.